% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar_plot}
\alias{calendar_plot}
\title{Retrieve personal data Github-like calendar plots}
\usage{
calendar_plot(data = calendar_data(), font = "Atkinson Hyperlegible")
}
\arguments{
\item{data}{tibble to draw the calendar}

\item{font}{font name}

\item{from}{Character of start date}

\item{to}{Character of enddate}

\item{github_user}{Github user name}

\item{github_repos}{A character vector of Github respositories (projects)}

\item{post_path}{Character of local path to search for posts}
}
\description{
Gather data from github contributions, repositories, posts and combine them
into an interactive calendar plot.
}
