% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_blocks.R
\name{code_blocks}
\alias{code_blocks}
\title{Convert scripts into R Markdown code blocks}
\usage{
code_blocks(dirs, exclude = NULL)
}
\arguments{
\item{dirs}{the directory containing scripts}
}
\value{

}
\description{
intended to use with chunk option results='asis'
}
\examples{
# convert multiple scripts inside directories
# code_blocks(c(dir1, dir2, ...))
# convert single script
# code_block(file)
}
