% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghutils.R
\name{gh_raw_url}
\alias{gh_raw_url}
\title{Utilities functions for working with github}
\usage{
gh_raw_url(repo, file, branch = "main", lfs = FALSE)
}
\arguments{
\item{repo}{user/repo}

\item{file}{file}

\item{branch}{branch}

\item{lfs}{if uses lfs storage}
}
\value{
raw file url for downloading
}
\description{
Utilities functions for working with github
}
\examples{
url <- gh_raw_url("qiushiyan/blog-data", "codes.csv", lfs = TRUE)
}
