% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_write.R
\name{group_write}
\alias{group_write}
\title{Divide a data frame into groups and write multiple csv files}
\usage{
group_write(df, ..., dir = "")
}
\arguments{
\item{df}{a data frame}

\item{...}{grouping variables}

\item{dir}{output directory. By default, the output files are stored under the current working directory}
}
\value{
\code{group_write} returns the input data frame invisibly.
}
\description{
\code{group_write} splits a data frame according to (multiple) grouping variables, write out each
part with file name indicating specific groups.
}
\examples{
library(ymisc)
group_write(iris, Species)
list.files(pattern = "\\\\.csv$")

# Currently, you need to convert `df` into a data frame outside this function
# before passing the argument.
# For example, `group_write(as.data.frame(Titanic), Sex, Age)` does not work
}
