% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long}
\alias{long}
\title{Long Format Sample Data}
\format{
A long format data frame. \describe{The regional brain measures from left and right hemisphere, attention check, and ID}
}
\usage{
long
}
\description{
Long Format Sample Data
}
\keyword{datasets}
