% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Wide Format Sample Data}
\format{
A wide format data frame. \describe{The regional brain measures from left and right hemisphere, attention check, and ID.}
}
\usage{
sample_data
}
\description{
Wide Format Sample Data
}
\keyword{datasets}
