% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold.R
\name{scaffold}
\alias{scaffold}
\alias{scaffold_widget}
\alias{scaffold_quarto_book}
\alias{scaffold_bookdown_book}
\title{Scaffold directories}
\usage{
scaffold(
  dir,
  path,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive(),
  is_package = FALSE
)

scaffold_widget(...)

scaffold_quarto_book(...)

scaffold_bookdown_book(...)
}
\arguments{
\item{dir}{directory to copy, relative to to inst/}

\item{path}{project path}

\item{rstudio}{if rstudio is available}

\item{open}{if rstudio is open}
}
\value{
the new project path
}
\description{
Similar to \code{packer::scaffold_widget()} but replace javascript with typescript, yarn with pnpm and webpack with esbuild
}
\examples{
\dontrun{
scaffold_widget("~/mywidget")
}
}
