% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_bookdown.R
\name{scaffold_bookdown}
\alias{scaffold_bookdown}
\title{Scaffold bookdown project with bs4 theme}
\usage{
scaffold_bookdown(
  path,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive(),
  gh_action = TRUE
)
}
\arguments{
\item{path}{project path}

\item{rstudio}{if rstudio is available}

\item{open}{if rstudio is open}

\item{gh_action}{whether to use a github action to deploy the book}
}
\value{
the new project path
}
\description{
\code{scaffold_bookdown()} generates a \code{bs4_book} format bookdown project.
Need to run \code{renv::init()} in the new project to use the action.
}
\examples{
\dontrun{scaffold_bookdown("~/mypkg")}
}
