% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_file_enc.R
\name{set_file_enc}
\alias{set_file_enc}
\title{Convert file encodings}
\usage{
set_file_enc(path, from = "", to = "UTF-8")
}
\arguments{
\item{path}{Either a path to a file, or a directory under which all files are to be converted}

\item{from}{A character string describing the current encoding. By default, uses the encoding detected by
\code{readr::guess_encoding} with highest probability}

\item{to}{A character string describing the target encoding, default to UTF-8.}
}
\description{
\code{set_file_enc} converts files between encodings. After converting, it is suggested using import functions from \code{readr},
i.e. \code{read_csv}, \code{read_lines}, etc.
}
