% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_knitr_options.R
\name{set_knitr_optons}
\alias{set_knitr_optons}
\title{set common knitr options}
\usage{
set_knitr_optons(...)
}
\arguments{
\item{...}{additional named arguments passed to knitr::opts_chunk$set()}
}
\description{
set common knitr options
}
