% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bookdown.R
\name{use_bookdown}
\alias{use_bookdown}
\title{Create a bs4_book bookdown project}
\usage{
use_bookdown(
  proj_name,
  dir = ".",
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{proj_name}{project name}

\item{dir}{parent directory of the project}

\item{rstudio}{if rstudio is available}

\item{open}{if rstudio is open}
}
\value{

}
\description{
Create a bs4_book bookdown project
}
\examples{
\dontrun{use_project("example_project")}
}
