% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_blocks.R
\name{code_blocks}
\alias{code_blocks}
\alias{code_block}
\title{Convert scripts into R Markdown code blocks}
\usage{
code_blocks(dirs, exclude = NULL, print_tree = TRUE)
}
\arguments{
\item{dirs}{directories containing scripts}

\item{exclude}{files or directories to exclude}

\item{print_tree}{whether file structure should be printed}
}
\value{
file contents
}
\description{
intended to be used with chunk option results='asis'
}
\examples{
# convert multiple scripts inside directories
\dontrun{code_blocks(c(dir1, dir2, ...))}
# convert single script
\dontrun{code_block(file)}
}
