% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bookdown.R
\name{use_bookdown}
\alias{use_bookdown}
\alias{create_bookdown}
\title{Generate project structures}
\usage{
use_bookdown(
  path,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive(),
  gh_action = TRUE
)

create_bookdown(
  path,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive(),
  gh_action = TRUE
)
}
\arguments{
\item{path}{project path}

\item{rstudio}{if rstudio is available}

\item{open}{if rstudio is open}

\item{gh_action}{whether to use a github action to deploy the book}
}
\value{
the new project path
}
\description{
\code{use_bookdown()} generates a \code{bs4_book} format bookdown project.
Need to run \code{renv::init()} in the new project to use the action. \code{create_bookdown} is an alias.
}
\examples{
\dontrun{create_bookdown("my-awesome-book", "~/documents")}
}
