% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ymisc-package.R
\docType{package}
\name{ymisc-package}
\alias{ymisc}
\alias{ymisc-package}
\title{ymisc: Qiushi Yan's miscellaneous functions}
\description{
my miscellaneous functions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qiushiyan.github.io/ymisc}
}

}
\author{
\strong{Maintainer}: Qiushi Yan \email{qiushi.yann@gmail.com}

}
\keyword{internal}
